silvermoon: cpi reboot
Allows Chumby 8 shutdown and reboot commands to work correctly. Taken
from Chumby 2.6.28 kernel. For this to work, you have to set the baud
rate of ttyS2 to 115200 manually in userspace before calling reboot or
shutdown. Would probably be cleaner to find a better place to put this
than inside the serial driver.