drm/stm: ltdc: fix pinctrl recovery after sleep
When the encoder is disabled, the pinctrl goes to sleep state. If the
encoder is re-enabled after that, the pinctrl needs to go back to
default state. This wasn't happening after the pinctrl setup was moved
to ltdc_encoder_mode_set.

Without this fix, if the framebuffer is put to sleep and then awakened
(e.g. using /sys/class/graphics/fb0/blank), it doesn't recover properly.

Signed-off-by: Doug Brown <doug@schmorgal.com>
Fixes: f412af187ae1 ("drm/stm: ltdc: move pinctrl to encoder mode set")