libertas: hack for PXA168 SDIO controller
I don't know how safe this is to do on any SDIO controller. But on the
PXA168, we really need to ack the IRQ as soon as possible. Because
another one could come in, and if we're not looking for them by the time
they arrive, we lose the IRQ.

This hack ensures that we're ready to detect new interrupts immediately
after we have acknowledged the last one.

For some reason, it doesn't work perfectly unless I re-read the IRQ
status register in the libertas module after re-enabling host card
interrupts. This makes no sense to me, but it solves the problem and I
no longer lose host interrupts after doing that.