clk: mmp: add SDHC clocks for pxa168
There are 4 SDHC peripherals. In addition to separate clock enables for
each one, SDH0/1 and 2/3 share common bits, which are treated as a
separate clock.

The existing code was missing APMU defines for SDH2 and 3, and a clock
define for SDH3. The parent clock names were also backwards: according
to the datasheet, mux value 0 selects a 48 MHz clock and mux value 1
selects a 52 MHz clock.