mmc: sdhci-pxav2: add support for PXA168 V1 controller
This driver was mostly compatible with the PXA168's SDHC controller, but
was missing a few workarounds that had been proposed in the past to work
around some silicon issues. This commit adds a new compatible string for
the version 1 controller used in the PXA168.

SDH2 and SDH4 will encounter a data abort exception if you try to access
their SDHCI_HOST_VERSION registers normally as word reads. The
workaround is to access them as long reads starting at a lower address.

This controller also needs the NO_BUSY_IRQ and NO_HISPD_BIT quirks.