clk: mmp: pxa168: only control functional bit on PWM peripherals
The PWM peripherals have shared enable and reset bits for PWM1/2 and
PWM3/4. Don't control these from Linux. For now, the bootloader will
need to turn them on. The shared bits are quirky in their behavior as
well, so it will be more reliable for the bootloader to handle them.