pwm-pxa: enable clock before applying config
The clock has to be on before we can apply the config anyway, so there
is no need to turn the clock on, off, and back on again.

The PWM clocks on the PXA168 don't like being turned off and back on
quickly. Prior to this change, I observed that the PXA168 wouldn't turn
the clock back on properly if the clock had just been stopped. This
change works around that problem.