pwm-pxa: set duty cycle to 0 before disabling
Turning off the clock doesn't guarantee that the pin will be left in a
low state. Setting the duty cycle to 0 forces it to turn off.

Without this fix, the PWM peripheral on a PXA168 was observed to leave
the PWM pin high after attempting to turn it off, especially when the
duty cycle was set to 100%.