ASoC: pxa: pxa-ssp: setup DMA data in probe
Follow the pattern that a lot of other drivers use, and set up the data
in probe rather allocating and freeing on demand in startup/shutdown.

This fixes an issue where the DMA data doesn't exist yet in
pxa2xx_pcm_open, so the channel never gets requested and eventually you
end up with a NULL pointer access.