ASoC: pxa: pxa-ssp: add support for LEFT_J format
This support is sort of specialized for the use case I need, but it
seems to work well. It works similarly to I2S mode, except it doesn't
have a dummy start because the data starts at the same time as the
LRCLK toggle. This mode also automatically pads mono data so it can be
fed through the interface as stereo, so we don't require any clocking
changes based on how many channels there are.

The automatic padding won't work with 32-bit samples though. It's only
tested with 16-bit samples.