ASoC: pxa: pxa-ssp: add preliminary support for PXA168
The PXA168 mostly has all the same hardware, but the clocking situation
is different.

This is loosely based on a patch from Haojian Zhuang from 2010, but
tweaked for my needs in order to get the Chumby 8's audio working.

I'm not proud of the current clocking solution, but it works. This
should probably be done the "correct" way with the common clock
framework. This is currently coded under the assumption that MCLK will
be asked to be 256fs. It's also coded under the assumption that the bit
clock will be 1/8 of MCLK. This works for the Chumby 8, but this patch
likely needs to be refined and genericized before we can get this in
mainline.