drm/armada: add support for PXA168 LCD controller
The PXA168 shares mostly the same LCD controller as Dove, with a few
features missing. This adds support for it.

This commit also adds an additional peripheral clock that can be
assigned for gating the entire peripheral, so that it will be compatible
with how the rest of the PXA168 peripherals work with the common clk
framework.